+++ packages/typegpu/src/std/index.ts
@@ -1,0 +1,5 @@
+import * as functions from './numeric';
+
+export const std = {
+  ...functions,
+};
